% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_centroid}
\alias{get_centroid}
\title{Extract Centroid}
\usage{
get_centroid(wkt)
}
\arguments{
\item{wkt}{a character vector of WKT objects, represented as strings}
}
\value{
a data.frame of two columns, \code{lat} and \code{lng},
with each row containing the centroid from the corresponding wkt
object. In the case that the object is NA (or cannot be decoded)
the resulting values will also be NA
}
\description{
\code{get_centroid} identifies the 2D centroid
in a WKT object (or vector of WKT objects).
}
\examples{
get_centroid("POLYGON((2 1.3,2.4 1.7))")
#  lat lng
#1 1.3   2
}

