% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wicket.R
\docType{package}
\name{wicket}
\alias{wicket}
\alias{wicket-package}
\title{Well-Known Text utilities}
\description{
This package provides basic, standalone functions for WKT
manipulation, generation and validation, backed by C++ for improved speed.
}
\seealso{
\code{\link{validate_wkt}} for checking the validity of WKT objects,
\code{\link{sp_convert}} for turning SP objects \emph{into} WKT, and
\code{\link{wkt_bounding}} and \code{\link{bounding_wkt}} for bounding box
conversion in both directions.
}

