% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_wkt}
\alias{validate_wkt}
\title{Validate WKT objects}
\usage{
validate_wkt(x)
}
\arguments{
\item{x}{a character vector of WKT objects.}
}
\value{
a data.frame of two columns, \code{is_valid} (containing TRUE or FALSE values
for whether the WKT object is valid) and \code{comments} (containing any error messages
in the case that the WKT object cannot be validated). If the objects are simply NA,
both fields will contain NA.
}
\description{
\code{validate_wkt} takes a vector of WKT objects and validates that
they are parsable, returning a data.frame containing the status of each entry and
(in the case it cannot be parsed) any comments as to what, in particular, may be
wrong with it.
}

