% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wkt_coords}
\alias{wkt_coords}
\title{Extract Latitude and Longitude from WKT polygons}
\usage{
wkt_coords(wkt)
}
\arguments{
\item{wkt}{a character vector of WKT objects}
}
\value{
a list of data.frames, one per WKT object. In the event that the
WKT object cannot be decoded (or is not a POLYGON), a data.frame of NA values
will be returned.
}
\description{
\code{wkt_coords} extracts lat/long values from WKT polygons,
specifically the outer shell of those polygons (working on the assumption that
said outer edge is what you want).
}
\examples{
wkt_coords("POLYGON ((30 10, 40 40, 20 40, 10 20, 30 10))")
#[[1]]
#lon lat
# 1  30  10
# 2  40  40
# 3  20  40
# 4  10  20
# 5  30  10
}

